+++ src/index.mjs
@@ -1,118 +1,119 @@
 // @flow
 
-import ENode from './nodes/en.mjs'
-import Button from './nodes/button.mjs'
-import A from './nodes/a.mjs'
-import Article from './nodes/article.mjs'
-import Aside from './nodes/aside.mjs'
-import Div from './nodes/div.mjs'
-import Footer from './nodes/footer.mjs'
-import H1 from './nodes/h1.mjs'
-import H2 from './nodes/h2.mjs'
-import H3 from './nodes/h3.mjs'
-import H4 from './nodes/h4.mjs'
-import H5 from './nodes/h5.mjs'
-import H6 from './nodes/h6.mjs'
-import Header from './nodes/header.mjs'
-import I from './nodes/i.mjs'
-import Main from './nodes/main.mjs'
-import P from './nodes/p.mjs'
-import Section from './nodes/section.mjs'
-import Span from './nodes/span.mjs'
-import Ul from './nodes/ul.mjs'
-import Ol from './nodes/ol.mjs'
-import Dl from './nodes/dl.mjs'
-import Li from './nodes/li.mjs'
-import Dt from './nodes/dt.mjs'
-import Dd from './nodes/dd.mjs'
-import Img from './nodes/img.mjs'
-import Input from './nodes/input.mjs'
-import Nav from './nodes/nav.mjs'
-import Abbr from './nodes/abbr.mjs'
-import Address from './nodes/address.mjs'
-import Area from './nodes/area.mjs'
-import Audio from './nodes/audio.mjs'
-import B from './nodes/b.mjs'
-import Base from './nodes/base.mjs'
-import Bdi from './nodes/bdi.mjs'
-import Bdo from './nodes/bdo.mjs'
-import Blockquote from './nodes/blockquote.mjs'
-import Body from './nodes/body.mjs'
-import Br from './nodes/br.mjs'
-import Canvas from './nodes/canvas.mjs'
-import Caption from './nodes/caption.mjs'
-import Cite from './nodes/cite.mjs'
-import Code from './nodes/code.mjs'
-import Col from './nodes/col.mjs'
-import Colgroup from './nodes/colgroup.mjs'
-import Data from './nodes/data.mjs'
-import Datalist from './nodes/datalist.mjs'
-import Del from './nodes/del.mjs'
-import Details from './nodes/details.mjs'
-import Dfn from './nodes/dfn.mjs'
-import Dialog from './nodes/dialog.mjs'
-import Em from './nodes/em.mjs'
-import Embed from './nodes/embed.mjs'
-import Fieldset from './nodes/fieldset.mjs'
-import Figcaption from './nodes/figcaption.mjs'
-import Figure from './nodes/figure.mjs'
-import Form from './nodes/form.mjs'
-import Head from './nodes/head.mjs'
-import Hr from './nodes/hr.mjs'
-import Html from './nodes/html.mjs'
-import IFrame from './nodes/iframe.mjs'
-import Ins from './nodes/ins.mjs'
-import Kbd from './nodes/kbd.mjs'
-import Label from './nodes/label.mjs'
-import Legend from './nodes/legend.mjs'
-import Link from './nodes/link.mjs'
-import Map from './nodes/map.mjs'
-import Mark from './nodes/mark.mjs'
-import Meta from './nodes/meta.mjs'
-import Meter from './nodes/meter.mjs'
-import NoScript from './nodes/noscript.mjs'
-import Object from './nodes/object.mjs'
-import OptGroup from './nodes/optgroup.mjs'
-import Option from './nodes/option.mjs'
-import Output from './nodes/output.mjs'
-import Param from './nodes/param.mjs'
-import Picture from './nodes/picture.mjs'
-import Pre from './nodes/pre.mjs'
-import Progress from './nodes/progress.mjs'
-import Q from './nodes/q.mjs'
-import Rp from './nodes/rp.mjs'
-import Rt from './nodes/rt.mjs'
-import Rtc from './nodes/rtc.mjs'
-import Ruby from './nodes/ruby.mjs'
-import S from './nodes/s.mjs'
-import Samp from './nodes/samp.mjs'
-import Script from './nodes/script.mjs'
-import Select from './nodes/select.mjs'
-import Slot from './nodes/slot.mjs'
-import Small from './nodes/small.mjs'
-import Source from './nodes/source.mjs'
-import Strong from './nodes/strong.mjs'
-import Style from './nodes/style.mjs'
-import Sub from './nodes/sub.mjs'
-import Summary from './nodes/summary.mjs'
-import Sup from './nodes/sup.mjs'
-import Table from './nodes/table.mjs'
-import TBody from './nodes/tbody.mjs'
-import Td from './nodes/td.mjs'
-import Template from './nodes/template.mjs'
-import TextArea from './nodes/textarea.mjs'
-import TFoot from './nodes/tfoot.mjs'
-import Th from './nodes/th.mjs'
-import THead from './nodes/thead.mjs'
-import Time from './nodes/time.mjs'
-import Title from './nodes/title.mjs'
-import Tr from './nodes/tr.mjs'
-import Track from './nodes/track.mjs'
-import U from './nodes/u.mjs'
-import Var from './nodes/var.mjs'
-import Video from './nodes/video.mjs'
-import Wbr from './nodes/wbr.mjs'
-import Custom from './nodes/custom.mjs'
+import {
+  button,
+  a,
+  article,
+  aside,
+  div,
+  footer,
+  h1,
+  h2,
+  h3,
+  h4,
+  h5,
+  h6,
+  header,
+  i,
+  main,
+  p,
+  section,
+  span,
+  ul,
+  li,
+  ol,
+  dl,
+  dt,
+  dd,
+  img,
+  input,
+  nav,
+  abbr,
+  address,
+  area,
+  audio,
+  b,
+  base,
+  bdi,
+  bdo,
+  blockquote,
+  body,
+  br,
+  canvas,
+  caption,
+  cite,
+  code,
+  col,
+  colgroup,
+  data,
+  datalist,
+  del,
+  details,
+  dfn,
+  dialog,
+  em,
+  embed,
+  fieldset,
+  figcaption,
+  figure,
+  form,
+  head,
+  hr,
+  html,
+  iframe,
+  ins,
+  kbd,
+  label,
+  legend,
+  link,
+  map,
+  mark,
+  meta,
+  meter,
+  noscript,
+  object,
+  optgroup,
+  option,
+  output,
+  param,
+  picture,
+  pre,
+  progress,
+  q,
+  rp,
+  rt,
+  rtc,
+  ruby,
+  s,
+  samp,
+  script,
+  select,
+  slot,
+  small,
+  source,
+  strong,
+  style,
+  sub,
+  summary,
+  sup,
+  table,
+  tbody,
+  td,
+  template,
+  textarea,
+  tfoot,
+  th,
+  thead,
+  time,
+  title,
+  tr,
+  track,
+  u,
+  Var,
+  video,
+  wbr,
+} from './elements/elements.mjs'
+import { custom } from './elements/element_function.mjs'
 
 import { createState } from './state/state.mjs'
 
@@ -125,120 +126,118 @@ export {
   Router,
   Component,
   I18n,
-
-  // --- start of nodes ---
-  ENode,
-  Button,
-  A,
-  Article,
-  Aside,
-  Div,
-  Footer,
-  H1,
-  H2,
-  H3,
-  H4,
-  H5,
-  H6,
-  Header,
-  I,
-  Main,
-  P,
-  Section,
-  Span,
-  Ul,
-  Li,
-  Ol,
-  Dl,
-  Dt,
-  Dd,
-  Img,
-  Input,
-  Nav,
-  Abbr,
-  Address,
-  Area,
-  Audio,
-  B,
-  Base,
-  Bdi,
-  Bdo,
-  Blockquote,
-  Body,
-  Br,
-  Canvas,
-  Caption,
-  Cite,
-  Code,
-  Col,
-  Colgroup,
-  Data,
-  Datalist,
-  Del,
-  Details,
-  Dfn,
-  Dialog,
-  Em,
-  Embed,
-  Fieldset,
-  Figcaption,
-  Figure,
-  Form,
-  Head,
-  Hr,
-  Html,
-  IFrame,
-  Ins,
-  Kbd,
-  Label,
-  Legend,
-  Link,
-  Map,
-  Mark,
-  Meta,
-  Meter,
-  NoScript,
-  Object,
-  OptGroup,
-  Option,
-  Output,
-  Param,
-  Picture,
-  Pre,
-  Progress,
-  Q,
-  Rp,
-  Rt,
-  Rtc,
-  Ruby,
-  S,
-  Samp,
-  Script,
-  Select,
-  Slot,
-  Small,
-  Source,
-  Strong,
-  Style,
-  Sub,
-  Summary,
-  Sup,
-  Table,
-  TBody,
-  Td,
-  Template,
-  TextArea,
-  TFoot,
-  Th,
-  THead,
-  Time,
-  Title,
-  Tr,
-  Track,
-  U,
+  // --- start of elements ---
+  button,
+  a,
+  article,
+  aside,
+  div,
+  footer,
+  h1,
+  h2,
+  h3,
+  h4,
+  h5,
+  h6,
+  header,
+  i,
+  main,
+  p,
+  section,
+  span,
+  ul,
+  li,
+  ol,
+  dl,
+  dt,
+  dd,
+  img,
+  input,
+  nav,
+  abbr,
+  address,
+  area,
+  audio,
+  b,
+  base,
+  bdi,
+  bdo,
+  blockquote,
+  body,
+  br,
+  canvas,
+  caption,
+  cite,
+  code,
+  col,
+  colgroup,
+  data,
+  datalist,
+  del,
+  details,
+  dfn,
+  dialog,
+  em,
+  embed,
+  fieldset,
+  figcaption,
+  figure,
+  form,
+  head,
+  hr,
+  html,
+  iframe,
+  ins,
+  kbd,
+  label,
+  legend,
+  link,
+  map,
+  mark,
+  meta,
+  meter,
+  noscript,
+  object,
+  optgroup,
+  option,
+  output,
+  param,
+  picture,
+  pre,
+  progress,
+  q,
+  rp,
+  rt,
+  rtc,
+  ruby,
+  s,
+  samp,
+  script,
+  select,
+  slot,
+  small,
+  source,
+  strong,
+  style,
+  sub,
+  summary,
+  sup,
+  table,
+  tbody,
+  td,
+  template,
+  textarea,
+  tfoot,
+  th,
+  thead,
+  time,
+  title,
+  tr,
+  track,
+  u,
   Var,
-  Video,
-  Wbr,
-  Custom
-  // --- end of nodes ---
+  video,
+  wbr,
+  custom,
+  // --- end of elements ---
 }