+++ src/render.mjs
@@ -1,19 +1,14 @@
 // @flow
 
-import { attach } from './nodes/en.mjs'
 import { uid } from './utils/uid.mjs'
-import { diff } from './utils/dom.mjs'
+import { diff, convertToDom } from './utils/dom.mjs'
 
-import type ENode from './nodes/en.mjs'
-import type Component from './component/component.mjs'
+import type { ElementChildren } from './elements/element_function.mjs'
 
 /**
  * Render [ENode] node or nodes and its derivate nodes as element or elements.
  */
-export function render(
-  to: string,
-  nodes: string | ENode | Component | (string | ENode | Component)[]
-): void {
+export function render(to: string, nodes: ElementChildren): void {
   const toElement = document.querySelector(to)
 
   if (toElement) {
@@ -21,7 +16,7 @@ export function render(
 
     newToElement.dataset.rid = toElement.dataset.rid || `${uid()}`
 
-    attach(newToElement, nodes)
+    newToElement.append(...convertToDom(nodes))
 
     toElement.hasChildNodes()
       ? diff(toElement, newToElement)